refactor: make input min, max length as constants
