+++ src/constants/input.ts
@@ -1,0 +1,8 @@
+export const MIN_NICKNAME_LENGTH = 2
+export const MAX_NICKNAME_LENGTH = 6
+
+export const MIN_MAP_NAME_LENGTH = 2
+export const MAX_MAP_NAME_LENGTH = 8
+
+export const MIN_MAP_DESCRIPTION_LENGTH = 0
+export const MAX_MAP_DESCRIPTION_LENGTH = 30
