--- src/constants/nickname.ts
@@ -1,2 +1,0 @@
-export const MIN_NICKNAME_LENGTH = 2
-export const MAX_NICKNAME_LENGTH = 6
